requirements.txt
flask
spotipy
discord.py
python-dotenv
lyricsgenius
openai
requests